chore: deploy service on ecs
A Terraform manifest is provided that defines the infrastructure for deploying the service on ECS.

It's replicated twice behind a load balancer. We can adjust that setup if need be. It doesn't have
DNS or SSL configured at the moment.

For now, the container has been uploaded to my own personal namespace on Dockerhub; we can create an
`autonomi` organisation at some point.